+++ test/ssr/test_deno_render_chunks.rb
@@ -4,25 +4,25 @@ require 'test_helper'
 require 'tmpdir'
 
 module SSR
-  class TestDenoRenderStreamChunks < Minitest::Test
-    CHUNKED_BUNDLE = File.expand_path('../fixtures/chunked-stream-bundle.js', __dir__)
+  class TestDenoRenderChunks < Minitest::Test
+    CHUNKED_BUNDLE = File.expand_path('../fixtures/chunked-bundle.js', __dir__)
 
     def setup
       @bundle = SSR::Deno::Bundle.new(CHUNKED_BUNDLE)
     end
 
-    def test_render_stream_chunks_with_block_yields_chunks
+    def test_render_chunks_with_block_yields_chunks
       chunks = []
 
-      @bundle.render_stream_chunks({ data: { name: 'chunked' } }) { |chunk| chunks << chunk }
+      @bundle.render_chunks({ data: { name: 'chunked' } }) { |chunk| chunks << chunk }
 
       assert_equal '<html><body>', chunks[0]
       assert_equal '<h1>chunked</h1>', chunks[1]
       assert_equal '</body></html>', chunks[2]
     end
 
-    def test_render_stream_chunks_without_block_returns_enumerator
-      enum = @bundle.render_stream_chunks({ data: { name: 'enum' } })
+    def test_render_chunks_without_block_returns_enumerator
+      enum = @bundle.render_chunks({ data: { name: 'enum' } })
 
       assert_kind_of Enumerator, enum
       chunks = enum.to_a
@@ -32,46 +32,46 @@ module SSR
       assert_equal '</body></html>', chunks[2]
     end
 
-    def test_render_stream_chunks_with_raw_input
+    def test_render_chunks_with_raw_input
       json = { data: { name: 'raw' } }.to_json
       chunks = []
 
-      @bundle.render_stream_chunks(json, raw_input: true) { |chunk| chunks << chunk }
+      @bundle.render_chunks(json, raw_input: true) { |chunk| chunks << chunk }
 
       assert_equal '<h1>raw</h1>', chunks[1]
     end
 
-    def test_render_stream_chunks_raises_on_promise_rejection
+    def test_render_chunks_raises_on_promise_rejection
       bundle = with_reject_bundle
       error = assert_raises(SSR::Deno::RenderError) do
-        bundle.render_stream_chunks({}) { |_chunk| nil }
+        bundle.render_chunks({}) { |_chunk| nil }
       end
 
       assert_includes error.message, 'chunked-rejection'
     end
 
-    def test_render_stream_chunks_enumerator_raises_on_promise_rejection
+    def test_render_chunks_enumerator_raises_on_promise_rejection
       bundle = with_reject_bundle
       error = assert_raises(SSR::Deno::RenderError) do
-        bundle.render_stream_chunks({}).to_a
+        bundle.render_chunks({}).to_a
       end
 
       assert_includes error.message, 'chunked-rejection'
     end
 
-    def test_render_stream_chunks_raises_on_timeout
+    def test_render_chunks_raises_on_timeout
       bundle = with_hang_bundle
 
       assert_raises(SSR::Deno::RenderError) do
-        bundle.render_stream_chunks({}) { |_chunk| nil }
+        bundle.render_chunks({}) { |_chunk| nil }
       end
     end
 
-    def test_render_stream_chunks_auto_reload
+    def test_render_chunks_auto_reload
       @bundle.auto_reload = true
       chunks = []
 
-      @bundle.render_stream_chunks({ data: { name: 'reload' } }) { |chunk| chunks << chunk }
+      @bundle.render_chunks({ data: { name: 'reload' } }) { |chunk| chunks << chunk }
 
       assert_includes chunks[1], 'reload'
     end